[SGMR-X] 쓰기 트랜잭션 경계를 도메인별 Writer로 통일 (RunningWriter·CourseWriter)
- RunningCreationWriter → RunningWriter 개명, RunningCommandService의 수정·삭제
  트랜잭션(updateRunningName/updateRunningPublicStatus/deleteRunnings)을 흡수
- RunningCommandService는 트랜잭션 밖 조율(가공·S3 업로드·VDOT)과 Writer 위임만 담당,
  진입점은 그대로 유지해 API 계층은 Writer의 존재를 모른다
- CourseService를 쓰기/조회로 분리: CourseWriter(save·updateCourse·deleteCourse·주인 구독 조율)
  + CourseQueryService(조회 전용)
- 호출자 갱신: CourseFacade, PushEventListener, PacemakerValidator, 문서(docs/core 현행화,
  설계 문서 2종에 개명 매핑 노트 추가)
- 테스트 재배치: CourseServiceTest → CourseQueryServiceTest/CourseWriterTest 분리,
  CourseServiceUnitTest → CourseWriterUnitTest, 수정·삭제 내부 검증은 RunningWriterTest로 이동,
  RunningCommandServiceTest는 위임 검증으로 대체
