+++ src/test/java/soma/ghostrunner/domain/course/application/CourseQueryServiceTest.java
@@ -1,0 +1,194 @@
+package soma.ghostrunner.domain.course.application;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import soma.ghostrunner.IntegrationTestSupport;
+import soma.ghostrunner.domain.course.dao.CourseRepository;
+import soma.ghostrunner.domain.course.domain.Course;
+import soma.ghostrunner.domain.course.dto.CourseSearchFilterDto;
+import soma.ghostrunner.domain.course.dto.CoursePreviewDto;
+import soma.ghostrunner.domain.course.enums.CourseSortType;
+import soma.ghostrunner.domain.member.infra.dao.MemberRepository;
+import soma.ghostrunner.domain.member.domain.Member;
+
+import java.util.List;
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+
+class CourseQueryServiceTest extends IntegrationTestSupport {
+
+    @Autowired private CourseQueryService courseQueryService;
+    @Autowired private CourseRepository courseRepository;
+    @Autowired private MemberRepository memberRepository;
+
+    private Member dummyMember;
+    private final double LAT = 37.54324;
+    private final double LNG = 126.94979;
+
+    @BeforeEach
+    void setUp() {
+        dummyMember = Member.of("카리나",  "karina.png");
+        memberRepository.save(dummyMember);
+    }
+
+
+    @DisplayName("주어진 위경도 반경 내의 공개된 코스 목록을 정상적으로 조회한다.")
+    @Test
+    void findNearbyCourses() {
+        // given
+        Course courseNearby1 = createPublicCourse("코스 1", LAT, LNG);
+        Course courseNearby2 = createPublicCourse("코스 2", LAT + 0.001, LNG - 0.001);
+        Course courseFar = createPublicCourse("먼 코스", LAT + 1, LNG + 1);
+        courseRepository.saveAll(List.of(courseNearby1, courseNearby2, courseFar));
+
+        // when
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(LAT, LNG, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        // - course1, 2는 조회되고, course3은 조회되지 않는다
+        Assertions.assertThat(courses).hasSize(2)
+                .extracting("name")
+                .containsExactlyInAnyOrder(courseNearby1.getName(), courseNearby2.getName());
+    }
+
+    @DisplayName("코스 조회 시 검색 반경이 0일 경우 정확히 검색한 지점의 코스만 조회한다.")
+    @Test
+    void findNearbyCourses_ZeroRadius() {
+        // given
+        Course course1 = createPublicCourse("코스 1", LAT, LNG);
+        Course course2 = createPublicCourse("코스 2", LAT + 0.001, LNG - 0.001);
+        courseRepository.saveAll(List.of(course1, course2));
+
+        // when
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(LAT, LNG, 0, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        Assertions.assertThat(courses).hasSize(1)
+                .extracting("name")
+                .containsExactly(course1.getName());
+    }
+
+    @DisplayName("코스가 비공개 상태인 경우 조회 시 반경 내에 있어도 조회할 수 없다.")
+    @Test
+    void findNearbyCourses_IsPublicFalse() {
+        // given
+        Course publicCourse = createPublicCourse("나를 찾아줘", LAT, LNG);
+        Course privateCourse = createPrivateCourse("나를 찾지마", LAT, LNG);
+        courseRepository.saveAll(List.of(publicCourse, privateCourse));
+
+        // when
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(LAT, LNG, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        Assertions.assertThat(courses).hasSize(1);
+        Assertions.assertThat(courses.get(0).name()).isEqualTo(publicCourse.getName());
+    }
+
+    @DisplayName("본초자오선 (경도 0도) 근처에서 코스를 검색하더라도 올바르게 조회할 수 있다.")
+    @Test
+    void searchCourse_PrimeMeridian() {
+        // given
+        Course courseEast = createPublicCourse("동경 코스", LAT, 0.001);
+        Course courseWest = createPublicCourse("서경 코스", LAT, -0.001);
+        Course courseFar = createPublicCourse("먼 코스", LAT, 0.1);
+        courseRepository.saveAll(List.of(courseEast, courseWest, courseFar));
+
+        // when
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(LAT, 0d, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        // 동경, 서경 코스는 모두 조회되고, 멀리 있는 코스는 조회되지 않아야 함
+        Assertions.assertThat(courses).hasSize(2)
+                .extracting("name")
+                .containsExactlyInAnyOrder(courseEast.getName(), courseWest.getName());
+    }
+
+    @DisplayName("날짜 변경선 (경도 180도) 근처에서 코스를 검색하더라도 올바르게 조회할 수 있다.")
+    @Test
+    void findNearbyCourses_DateLine() {
+        // given
+        // 날짜 변경선 근처인 동경(양수) 끝과 서경(음수) 끝에 코스를 생성
+        Course courseEast = createPublicCourse("동경 끝 코스", LAT, 179.999);
+        Course courseWest = createPublicCourse("서경 끝 코스", LAT, -179.999);
+        Course courseFar = createPublicCourse("먼 코스", LAT, 0);
+        courseRepository.saveAll(List.of(courseEast, courseWest, courseFar));
+
+        // when
+        // 동경 179.9985도 지점에서 반경 1km 내 코스 검색
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(LAT, 179.9985, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        // 동경 끝, 서경 끝 코스는 모두 조회되고, 멀리 있는 코스는 조회되지 않아야 함
+        // todo
+//        Assertions.assertThat(courses).hasSize(2)
+//                .extracting("name")
+//                .containsExactlyInAnyOrder(courseEast.getName(), courseWest.getName());
+    }
+
+    @DisplayName("본초자오선 기준으로 위경도가 이루는 4분면 어디에서도 코스를 조회할 수 있다.")
+    @ParameterizedTest(name = "[{index}] {0}에서 코스를 검색하면 올바른 코스가 조회된다.")
+    @MethodSource("locationQuadrants")
+    void findNearbyCourses_quadrants(String name, double lat, double lng) {
+        // given
+        Course courseNearby = createPublicCourse("근처 코스", lat + 0.001, lng + 0.001);
+        Course courseFar = createPublicCourse("먼 코스", lat + 1.0, lng + 1.0);
+        courseRepository.saveAll(List.of(courseNearby, courseFar));
+
+        // when
+        List<CoursePreviewDto> courses = courseQueryService.findNearbyCourses(lat, lng, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
+
+        // then
+        assertThat(courses).hasSize(1)
+                .extracting("name")
+                .containsExactly(courseNearby.getName());
+    }
+
+    private static Stream<Arguments> locationQuadrants() {
+        return Stream.of(
+                Arguments.of("북동쪽 (1사분면)", 37.5, 127.5),
+                Arguments.of("북서쪽 (2사분면)", 37.5, -127.5),
+                Arguments.of("남서쪽 (3사분면)", -36.5, 127.5),
+                Arguments.of("남동쪽 (4사분면)", -36.5, -127.5)
+        );
+    }
+
+
+    @DisplayName("코스의 id를 기반으로 코스 상세 정보를 조회할 수 있다.")
+    @Test
+    void getCourseDetail() {
+        // given
+        Course course = createPublicCourse("course", LAT, LNG);
+        Course savedCourse = courseRepository.save(course);
+
+        // when
+        Course foundCourse = courseQueryService.findCourseById(savedCourse.getId());
+
+        // then
+        Assertions.assertThat(foundCourse.getId()).isEqualTo(savedCourse.getId());
+    }
+
+    // --- Helper Methods ---
+    private Course createPublicCourse(String name, double lat, double lng) {
+        return createCourse(name, lat, lng, true);
+    }
+
+    private Course createPrivateCourse(String name, double lat, double lng) {
+        return createCourse(name, lat, lng, false);
+    }
+
+    private Course createCourse(String name, double lat, double lng, boolean isPublic) {
+        Course course = Course.of(dummyMember, 0d, 0d, 0d, 0d, lat, lng, "url", "url", "url");
+        course.setName(name);
+        course.setIsPublic(isPublic);
+        return course;
+    }
+
+}
