+++ src/test/java/soma/ghostrunner/domain/course/application/CourseWriterTest.java
@@ -4,19 +4,13 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import soma.ghostrunner.IntegrationTestSupport;
 import soma.ghostrunner.domain.course.dao.CourseReadModelRepository;
 import soma.ghostrunner.domain.course.dao.CourseRepository;
 import soma.ghostrunner.domain.course.domain.Course;
 import soma.ghostrunner.domain.course.domain.CourseReadModel;
-import soma.ghostrunner.domain.course.dto.CourseSearchFilterDto;
-import soma.ghostrunner.domain.course.dto.CoursePreviewDto;
 import soma.ghostrunner.domain.course.dto.request.CoursePatchRequest;
-import soma.ghostrunner.domain.course.enums.CourseSortType;
 import soma.ghostrunner.domain.course.exception.CourseNameNotValidException;
 import soma.ghostrunner.domain.member.infra.dao.MemberRepository;
 import soma.ghostrunner.domain.member.domain.Member;
@@ -27,15 +21,17 @@ import soma.ghostrunner.domain.running.infra.persistence.RunningRepository;
 
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static soma.ghostrunner.domain.course.dto.request.CoursePatchRequest.UpdatedAttr.*;
 
+/**
+ * 코스 쓰기 트랜잭션 경계({@link CourseWriter})의 통합 테스트 — 수정·삭제와 리드모델 동기화.
+ * (구 CourseServiceTest 의 쓰기 절반. 조회는 CourseQueryServiceTest 가 덮는다)
+ */
+class CourseWriterTest extends IntegrationTestSupport {
 
-class CourseServiceTest extends IntegrationTestSupport {
-
-    @Autowired private CourseService courseService;
+    @Autowired private CourseWriter courseWriter;
     @Autowired private CourseRepository courseRepository;
     @Autowired private MemberRepository memberRepository;
     @Autowired private RunningRepository runningRepository;
@@ -51,142 +47,6 @@ class CourseServiceTest extends IntegrationTestSupport {
         memberRepository.save(dummyMember);
     }
 
-
-    @DisplayName("주어진 위경도 반경 내의 공개된 코스 목록을 정상적으로 조회한다.")
-    @Test
-    void findNearbyCourses() {
-        // given
-        Course courseNearby1 = createPublicCourse("코스 1", LAT, LNG);
-        Course courseNearby2 = createPublicCourse("코스 2", LAT + 0.001, LNG - 0.001);
-        Course courseFar = createPublicCourse("먼 코스", LAT + 1, LNG + 1);
-        courseRepository.saveAll(List.of(courseNearby1, courseNearby2, courseFar));
-
-        // when
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(LAT, LNG, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        // - course1, 2는 조회되고, course3은 조회되지 않는다
-        Assertions.assertThat(courses).hasSize(2)
-                .extracting("name")
-                .containsExactlyInAnyOrder(courseNearby1.getName(), courseNearby2.getName());
-    }
-
-    @DisplayName("코스 조회 시 검색 반경이 0일 경우 정확히 검색한 지점의 코스만 조회한다.")
-    @Test
-    void findNearbyCourses_ZeroRadius() {
-        // given
-        Course course1 = createPublicCourse("코스 1", LAT, LNG);
-        Course course2 = createPublicCourse("코스 2", LAT + 0.001, LNG - 0.001);
-        courseRepository.saveAll(List.of(course1, course2));
-
-        // when
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(LAT, LNG, 0, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        Assertions.assertThat(courses).hasSize(1)
-                .extracting("name")
-                .containsExactly(course1.getName());
-    }
-
-    @DisplayName("코스가 비공개 상태인 경우 조회 시 반경 내에 있어도 조회할 수 없다.")
-    @Test
-    void findNearbyCourses_IsPublicFalse() {
-        // given
-        Course publicCourse = createPublicCourse("나를 찾아줘", LAT, LNG);
-        Course privateCourse = createPrivateCourse("나를 찾지마", LAT, LNG);
-        courseRepository.saveAll(List.of(publicCourse, privateCourse));
-
-        // when
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(LAT, LNG, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        Assertions.assertThat(courses).hasSize(1);
-        Assertions.assertThat(courses.get(0).name()).isEqualTo(publicCourse.getName());
-    }
-
-    @DisplayName("본초자오선 (경도 0도) 근처에서 코스를 검색하더라도 올바르게 조회할 수 있다.")
-    @Test
-    void searchCourse_PrimeMeridian() {
-        // given
-        Course courseEast = createPublicCourse("동경 코스", LAT, 0.001);
-        Course courseWest = createPublicCourse("서경 코스", LAT, -0.001);
-        Course courseFar = createPublicCourse("먼 코스", LAT, 0.1);
-        courseRepository.saveAll(List.of(courseEast, courseWest, courseFar));
-
-        // when
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(LAT, 0d, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        // 동경, 서경 코스는 모두 조회되고, 멀리 있는 코스는 조회되지 않아야 함
-        Assertions.assertThat(courses).hasSize(2)
-                .extracting("name")
-                .containsExactlyInAnyOrder(courseEast.getName(), courseWest.getName());
-    }
-
-    @DisplayName("날짜 변경선 (경도 180도) 근처에서 코스를 검색하더라도 올바르게 조회할 수 있다.")
-    @Test
-    void findNearbyCourses_DateLine() {
-        // given
-        // 날짜 변경선 근처인 동경(양수) 끝과 서경(음수) 끝에 코스를 생성
-        Course courseEast = createPublicCourse("동경 끝 코스", LAT, 179.999);
-        Course courseWest = createPublicCourse("서경 끝 코스", LAT, -179.999);
-        Course courseFar = createPublicCourse("먼 코스", LAT, 0);
-        courseRepository.saveAll(List.of(courseEast, courseWest, courseFar));
-
-        // when
-        // 동경 179.9985도 지점에서 반경 1km 내 코스 검색
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(LAT, 179.9985, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        // 동경 끝, 서경 끝 코스는 모두 조회되고, 멀리 있는 코스는 조회되지 않아야 함
-        // todo
-//        Assertions.assertThat(courses).hasSize(2)
-//                .extracting("name")
-//                .containsExactlyInAnyOrder(courseEast.getName(), courseWest.getName());
-    }
-
-    @DisplayName("본초자오선 기준으로 위경도가 이루는 4분면 어디에서도 코스를 조회할 수 있다.")
-    @ParameterizedTest(name = "[{index}] {0}에서 코스를 검색하면 올바른 코스가 조회된다.")
-    @MethodSource("locationQuadrants")
-    void findNearbyCourses_quadrants(String name, double lat, double lng) {
-        // given
-        Course courseNearby = createPublicCourse("근처 코스", lat + 0.001, lng + 0.001);
-        Course courseFar = createPublicCourse("먼 코스", lat + 1.0, lng + 1.0);
-        courseRepository.saveAll(List.of(courseNearby, courseFar));
-
-        // when
-        List<CoursePreviewDto> courses = courseService.findNearbyCourses(lat, lng, 1000, CourseSortType.DISTANCE, CourseSearchFilterDto.of(), dummyMember.getId());
-
-        // then
-        assertThat(courses).hasSize(1)
-                .extracting("name")
-                .containsExactly(courseNearby.getName());
-    }
-
-    private static Stream<Arguments> locationQuadrants() {
-        return Stream.of(
-                Arguments.of("북동쪽 (1사분면)", 37.5, 127.5),
-                Arguments.of("북서쪽 (2사분면)", 37.5, -127.5),
-                Arguments.of("남서쪽 (3사분면)", -36.5, 127.5),
-                Arguments.of("남동쪽 (4사분면)", -36.5, -127.5)
-        );
-    }
-
-
-    @DisplayName("코스의 id를 기반으로 코스 상세 정보를 조회할 수 있다.")
-    @Test
-    void getCourseDetail() {
-        // given
-        Course course = createPublicCourse("course", LAT, LNG);
-        Course savedCourse = courseRepository.save(course);
-
-        // when
-        Course foundCourse = courseService.findCourseById(savedCourse.getId());
-
-        // then
-        Assertions.assertThat(foundCourse.getId()).isEqualTo(savedCourse.getId());
-    }
-
     @DisplayName("코스의 제목과 공개 여부를 수정할 수 있다.")
     @Test
     void updateCourse_Success() {
@@ -196,7 +56,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest("바꿨다", true, Set.of(NAME, IS_PUBLIC));
 
         // when
-        courseService.updateCourse(id, request, dummyMember.getUuid());
+        courseWriter.updateCourse(id, request, dummyMember.getUuid());
 
         // then
         Course course = courseRepository.findById(id).orElseThrow();
@@ -213,7 +73,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest("바꿨다", null, Set.of(NAME));
 
         // when
-        courseService.updateCourse(id, request, dummyMember.getUuid());
+        courseWriter.updateCourse(id, request, dummyMember.getUuid());
 
         // then
         Course course = courseRepository.findById(id).orElseThrow();
@@ -230,7 +90,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(" ", null, Set.of(NAME));
 
         // when & then
-        Assertions.assertThatThrownBy(() -> courseService.updateCourse(id, request, dummyMember.getUuid()))
+        Assertions.assertThatThrownBy(() -> courseWriter.updateCourse(id, request, dummyMember.getUuid()))
                 .isInstanceOf(CourseNameNotValidException.class)
                 .hasMessage("invalid course name");
     }
@@ -244,7 +104,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, false, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(id, request, dummyMember.getUuid());
+        courseWriter.updateCourse(id, request, dummyMember.getUuid());
 
         // then - 등록 해제 성공
         Course course = courseRepository.findById(id).orElseThrow();
@@ -260,7 +120,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, true, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(id, request, dummyMember.getUuid());
+        courseWriter.updateCourse(id, request, dummyMember.getUuid());
 
         // then - 여전히 공개 상태 유지 (멱등성)
         Course course = courseRepository.findById(id).orElseThrow();
@@ -276,7 +136,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, false, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(id, request, dummyMember.getUuid());
+        courseWriter.updateCourse(id, request, dummyMember.getUuid());
 
         // then - 여전히 비공개 상태 유지 (멱등성)
         Course course = courseRepository.findById(id).orElseThrow();
@@ -291,7 +151,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         Long id = courseRepository.save(course).getId();
 
         // when
-        courseService.deleteCourse(id, dummyMember.getUuid());
+        courseWriter.deleteCourse(id, dummyMember.getUuid());
 
         // then
         Assertions.assertThat(courseRepository.findById(id)).isNotPresent();
@@ -330,7 +190,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, true, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(course.getId(), request, dummyMember.getUuid());
+        courseWriter.updateCourse(course.getId(), request, dummyMember.getUuid());
 
         // then
         CourseReadModel readModel = readModelRepository.findByCourseId(course.getId())
@@ -358,11 +218,11 @@ class CourseServiceTest extends IntegrationTestSupport {
         runningRepository.save(run);
 
         // 처음 등록
-        courseService.updateCourse(course.getId(),
+        courseWriter.updateCourse(course.getId(),
                 new CoursePatchRequest(null, true, Set.of(IS_PUBLIC)), dummyMember.getUuid());
 
         // 등록 해제
-        courseService.updateCourse(course.getId(),
+        courseWriter.updateCourse(course.getId(),
                 new CoursePatchRequest(null, false, Set.of(IS_PUBLIC)), dummyMember.getUuid());
 
         CourseReadModel readModelAfterUnregister = readModelRepository.findByCourseId(course.getId())
@@ -370,7 +230,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         assertThat(readModelAfterUnregister.getIsPublic()).isFalse();
 
         // when - 재등록
-        courseService.updateCourse(course.getId(),
+        courseWriter.updateCourse(course.getId(),
                 new CoursePatchRequest(null, true, Set.of(IS_PUBLIC)), dummyMember.getUuid());
 
         // then
@@ -397,7 +257,7 @@ class CourseServiceTest extends IntegrationTestSupport {
 
         // when & then
         Assertions.assertThatThrownBy(() ->
-                courseService.updateCourse(course.getId(), request, dummyMember.getUuid()))
+                courseWriter.updateCourse(course.getId(), request, dummyMember.getUuid()))
                 .isInstanceOf(CourseNameNotValidException.class);
     }
 
@@ -416,7 +276,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, true, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(course.getId(), request, dummyMember.getUuid());
+        courseWriter.updateCourse(course.getId(), request, dummyMember.getUuid());
 
         // then
         CourseReadModel readModel = readModelRepository.findByCourseId(course.getId())
@@ -434,13 +294,13 @@ class CourseServiceTest extends IntegrationTestSupport {
         courseRepository.save(course);
         runningRepository.save(createRunning(course, dummyMember, 100L, false));
 
-        courseService.updateCourse(course.getId(),
+        courseWriter.updateCourse(course.getId(),
                 new CoursePatchRequest(null, true, Set.of(IS_PUBLIC)), dummyMember.getUuid());
         assertThat(readModelRepository.findByCourseId(course.getId()).orElseThrow().getName())
                 .isEqualTo("옛 이름");
 
         // when
-        courseService.updateCourse(course.getId(),
+        courseWriter.updateCourse(course.getId(),
                 new CoursePatchRequest("새 이름", null, Set.of(NAME)), dummyMember.getUuid());
 
         // then
@@ -464,7 +324,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, true, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(course.getId(), request, dummyMember.getUuid());
+        courseWriter.updateCourse(course.getId(), request, dummyMember.getUuid());
 
         // then - 주인 기록만 채우는 것이 아니라 전체 러닝으로 초기화된다
         CourseReadModel readModel = readModelRepository.findByCourseId(course.getId())
@@ -491,7 +351,7 @@ class CourseServiceTest extends IntegrationTestSupport {
         CoursePatchRequest request = new CoursePatchRequest(null, true, Set.of(IS_PUBLIC));
 
         // when
-        courseService.updateCourse(course.getId(), request, dummyMember.getUuid());
+        courseWriter.updateCourse(course.getId(), request, dummyMember.getUuid());
 
         // then
         CourseReadModel readModel = readModelRepository.findByCourseId(course.getId())
